Contents: avoid double-URIdecoding paths
The paths coming in to /api/contents need to be URI decoded, because
jupyterlab's frontend URI encodes them.

The existing solution would  sometimes double-decode the filenames. For
example, `Contents.createCheckpoint()` would decode the path once, then
need call `this.get()` on the path, which decoded it again.

To prevent double-decoding, the URI decode is moved out of the Contents
class and into the route definitions.

This has a side effect of changing the Contents API to accept "regular"
paths, and not URI-encoded ones.

Previously where to access a file called "hello world.txt", a method
call would look like:

```
await contents.get("hello%20world.txt")
```

it now looks like:

```
await contents.get("hello world").txt
```

which I think seems more natural anyway. But it is a breaking change to
that API.

ci